Map part of Lab-1
Coordinator will send map task to a worker if asked.
If there are no more map task in chan, it will tell worker to wait for 30 seconds.

Still need to implement the health check part, so that the coordinator can recycle jobs from slow or died builder, and re-assign them.

But maybe should implement this after finish the shuffle and reduce part.